chore: renamed view
